requirements.txt
django==6.0.1
djangorestframework==3.14.0
django-rest-framework-jsonapi==4.0.0
drf-spectacular==0.29.0
drf-yasg==1.21.14
black==25.12.0
isort==7.0.0